fixed bug with batch sampling for recurrent ppo
